--- a/pkg/frontend/querymiddleware/limits_test.go
+++ b/pkg/frontend/querymiddleware/limits_test.go
@@ -707,3 +707,68 @@ func TestLimitedRoundTripper_OriginalRequestContextCancellation(t *testing.T) {
 	).RoundTrip(r)
 	require.NoError(t, err)
 }
+
+func BenchmarkLimitedParallelismRoundTripper(b *testing.B) {
+	maxParallelism := 10
+	workDuration := 20 * time.Millisecond
+	ctx := user.InjectOrgID(context.Background(), "test-org-id")
+
+	downstream := RoundTripFunc(func(_ *http.Request) (*http.Response, error) {
+		// Simulate some work
+		time.Sleep(workDuration)
+		return &http.Response{
+			Body: http.NoBody,
+		}, nil
+	})
+
+	codec := newTestPrometheusCodec()
+	r, err := codec.EncodeRequest(ctx, &PrometheusRangeQueryRequest{
+		Path:  "/api/v1/query_range",
+		Start: time.Now().Add(time.Hour).Unix(),
+		End:   util.TimeToMillis(time.Now()),
+		Step:  int64(1 * time.Second * time.Millisecond),
+		Query: `foo`,
+	})
+	require.Nil(b, err)
+
+	for _, concurrentRequestCount := range []int{1, 10, 100} {
+		for _, subRequestCount := range []int{1, 2, 5, 10, 20, 50, 100} {
+			tripper := newLimitedParallelismRoundTripper(downstream, codec, mockLimits{maxQueryParallelism: maxParallelism},
+				MiddlewareFunc(func(next Handler) Handler {
+					return HandlerFunc(func(c context.Context, _ Request) (Response, error) {
+						wg := sync.WaitGroup{}
+						for i := 0; i < subRequestCount; i++ {
+							wg.Add(1)
+							go func() {
+								defer wg.Done()
+								_, _ = next.Do(c, &PrometheusRangeQueryRequest{})
+							}()
+						}
+						wg.Wait()
+						return newEmptyPrometheusResponse(), nil
+					})
+				}),
+			)
+
+			b.Run(fmt.Sprintf("%v concurrent requests with %v sub requests each, max parallelism %v, sub request duration %v", concurrentRequestCount, subRequestCount, maxParallelism, workDuration.String()), func(b *testing.B) {
+				for i := 0; i < b.N; i++ {
+					wg := sync.WaitGroup{}
+
+					for i := 0; i < concurrentRequestCount; i++ {
+						wg.Add(1)
+						go func() {
+							defer wg.Done()
+							_, err := tripper.RoundTrip(r)
+
+							if err != nil {
+								require.NoError(b, err)
+							}
+						}()
+					}
+
+					wg.Wait()
+				}
+			})
+		}
+	}
+}