#### What this PR does

This PR improves the performance of `limitedParallelismRoundTripper` by reducing the number of goroutines spawned.

This change is based on https://github.com/grafana/loki/pull/9923.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated
- [n/a] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`