@charleskorn 2023-09-20
Note to reviewers: I'm not sure a changelog entry is necessary for this, but happy to add one if you disagree.