@charleskorn 2023-09-21
> Do we have any benchmark to run?

There's no existing benchmark, but I'll add one.